+++ src/Event/AfterApplicationInitializationEvent.php
@@ -3,14 +3,13 @@
 namespace Castor\Event;
 
 use Castor\Console\Application;
-use Castor\TaskDescriptor;
+use Castor\TaskDescriptorCollection;
 
 class AfterApplicationInitializationEvent
 {
     public function __construct(
         public readonly Application $application,
-        /** @var array<TaskDescriptor> $tasks */
-        public array &$tasks,
+        public TaskDescriptorCollection $taskDescriptorCollection,
     ) {
     }
 }