+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Not released yet
 
 * Add an error handler, and wire the logger to it so display deprecation notices
+* Add `guard_min_version()` function to ensure a minimum version of Castor is used
+* Edited the duration of update check from `60 days` to `24 hours`
 
 ## 0.11.1 (2024-01-11)